first 50 Fibonacci numbers, starting with 1 and 2 ,fix size

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,15 +6,14 @@
  */
 int main(void)
 {
-	double y[50] = {1, 2};
+	unsigned long y[50] = {1, 2};
 	int count;
 
 	printf("1, 2, ");
 	for (count = 2 ; count < 51 ; count++)
 	{
 		y[count] = y[count - 1] + y[count - 2];
-		/*%.01f gor double means no decimal places*/
-		printf("%.01f", y[count]);
+		printf("%lu", y[count]);
 		if (count < 49)
 			printf(", ");
 	}
